fix(binder): hoist var declarations from with-statement bodies (TS2304)
var inside a with(...) body is function-scoped and hoisted by tsc
(Annex B). collect_hoisted_declarations_impl recursed into
if/while/for/try/switch/labeled bodies but not with, so a var declared
inside a with body was never collected -> a later reference or
export { name } resolved to nothing and tsz emitted a spurious TS2304.

Add a WITH_STATEMENT arm mirroring the sibling LABELED_STATEMENT arm
(recurse into then_statement via collect_hoisted_from_node, which
dispatches BLOCK vs single-statement). This is a coverage gap, not a
new concept.

To keep binding.rs under its size ratchet, the hoisting method group
is extracted verbatim into a sibling nodes/hoisting.rs impl block
(binding.rs 2027 -> 1767); the ratchet is tightened 2018 -> 1767 so the
extraction cannot silently regrow.

Flips topLevelVarHoistingCommonJS.ts (exercises every hoisting
container); 0 regressions across a 202-test hoisting/var/with/export net.
